refactor(plugin-hubtype-analytics): rename EventType.flow to EventType.botevent, and remove classic from intent-classic
