+++ packages/botonic-plugin-hubtype-analytics/src/event-models/ht-event-flow.ts
@@ -16,7 +16,7 @@ export class HtEventFlow extends HtEvent {
 
   constructor(event: EventFlow, requestData: RequestData) {
     super(event, requestData)
-    this.type = EventType.flow
+    this.type = EventType.botevent
     this.data.flow_thread_id = event.data.flowThreadId // This value is managed by the flow builder plugin, stored in the session and updated every time the content connected to the conversation start is displayed
     this.data.flow_id = event.data.flowId
     this.data.flow_name = event.data.flowName
