+++ packages/botonic-plugin-hubtype-analytics/src/event-models/ht-event-intent.ts
