+++ packages/botonic-plugin-hubtype-analytics/src/event-models/ht-event-keyword.ts
@@ -14,7 +14,7 @@ export class HtEventKeyword extends HtEvent {
 
   constructor(event: EventKeyword, requestData: RequestData) {
     super(event, requestData)
-    this.type = EventType.flow
+    this.type = EventType.botevent
     this.data.nlu_keyword_id = event.data.nluKeywordId
     this.data.nlu_keyword_name = event.data.nluKeywordName
     this.data.nlu_keyword_is_regex = event.data.nluKeywordIsRegex || false
