+++ packages/botonic-plugin-hubtype-analytics/src/event-models/ht-event-knowledge-base.ts
@@ -8,8 +8,7 @@ import { HtEvent } from './ht-event'
 
 interface EventDataKnowledgeBase {
   action: EventAction.knowledgebase
-  knowledgebase_id: string
-  knowledgebase_fail_reason: string
+  knowledgebase_fail_reason?: string
   knowledgebase_sources_ids: string[]
   knowledgebase_chunks_ids: string[]
 }
@@ -19,8 +18,7 @@ export class HtEventKnowledgeBase extends HtEvent {
 
   constructor(event: EventKnowledgeBase, requestData: RequestData) {
     super(event, requestData)
-    this.type = EventType.flow
-    this.data.knowledgebase_id = event.data.knowledgebaseId
+    this.type = EventType.botevent
     this.data.knowledgebase_fail_reason = event.data.knowledgebaseFailReason
     this.data.knowledgebase_sources_ids = event.data.knowledgebaseSourcesIds
     this.data.knowledgebase_chunks_ids = event.data.knowledgebaseChunksIds
