+++ packages/botonic-plugin-hubtype-analytics/src/event-models/index.ts
@@ -3,7 +3,7 @@ export { HtEventFallback } from './ht-event-fallback'
 export { HtEventFeedback } from './ht-event-feedback'
 export { HtEventFlow } from './ht-event-flow'
 export { HtEventHandoff } from './ht-event-handoff'
-export { HtEventIntentClassic } from './ht-event-intent-classic'
+export { HtEventIntent } from './ht-event-intent'
 export { HtEventIntentSmart } from './ht-event-intent-smart'
 export { HtEventKeyword } from './ht-event-keyword'
 export { HtEventKnowledgeBase } from './ht-event-knowledge-base'
