+++ packages/botonic-plugin-hubtype-analytics/tests/event-flow.test.ts
@@ -33,7 +33,7 @@ describe('Create flow event', () => {
           flow_node_content_id: 'flowNodeContentIdTest',
           flow_node_is_meaningful: false,
         },
-        type: EventType.flow,
+        type: EventType.botevent,
       })
     )
   })
