+++ packages/botonic-plugin-hubtype-analytics/tests/event-handoff.test.ts
@@ -33,7 +33,7 @@ describe('Create handoff events', () => {
           handoff_is_available_agent: true,
           handoff_is_threshold_reached: false,
         },
-        type: EventType.flow,
+        type: EventType.botevent,
       })
     )
   })
@@ -66,7 +66,7 @@ describe('Create handoff events', () => {
           handoff_is_available_agent: false,
           handoff_is_threshold_reached: false,
         },
-        type: EventType.flow,
+        type: EventType.botevent,
       })
     )
   })
@@ -93,7 +93,7 @@ describe('Create handoff events', () => {
           handoff_queue_id: 'handoffQueueIdTest',
           handoff_queue_name: 'handoffQueueNameTest',
         },
-        type: EventType.flow,
+        type: EventType.botevent,
       })
     )
   })
